Revert to shared pointer-based sortable; harden for touch
Drop SortableJS and go back to the unified pointer-drag controller (indicator line +
content-swap for bullets, row move + PATCH for plain lists) so the sequence pages list
and the bullets editor feel the same. Pointer Events already cover touch; add a
pointercancel abort and keep touch-action:none on the grip so a touch-drag won't be
hijacked into a scroll. Verified the touch path reorders and persists.
